+++ InternetArchive.NET.Tests/JsonConverterTests.cs
@@ -11,12 +11,12 @@ public class JsonConverterTests
         var response = new Metadata.ReadResponse();
         var json = JsonSerializer.Serialize(response);
 
-        using var test = JsonSerializer.Deserialize<Metadata.ReadResponse>(json);
+        var test = JsonSerializer.Deserialize<Metadata.ReadResponse>(json);
 
         Assert.IsNotNull(test);
         Assert.IsNull(test.WorkableServers);
 
-        using var test2 = JsonSerializer.Deserialize<Metadata.ReadResponse>
+        var test2 = JsonSerializer.Deserialize<Metadata.ReadResponse>
         (
             "{\"workable_servers\":\"1\"}"
         );
@@ -26,7 +26,7 @@ public class JsonConverterTests
         response.WorkableServers = new[] { "1", "2" };
 
         json = JsonSerializer.Serialize(response);
-        using var test3 = JsonSerializer.Deserialize<Metadata.ReadResponse>(json);
+        var test3 = JsonSerializer.Deserialize<Metadata.ReadResponse>(json);
 
         Assert.AreEqual(2, test3?.WorkableServers?.Count());
     }